added api to create inventory item

+++ src/migrations/20221017134213-create-inventory.js.js
@@ -22,7 +22,7 @@ module.exports = {
           },
         },
         amountAvailable: {
-          type: Sequelize.DataTypes.INTEGER,
+          type: Sequelize.DataTypes.FLOAT,
           allowNull: false,
         },
